Update index.html
Changed the Graphic Design and Website Design images to be able to trigger the slideshow. Added slideshow images to page. Added links to the JavaScript and CSS for the slideshow. (The slideshow was added tentatively since I don't know if we were going to make any changes to this page after all or not) Removed Google analytics script from the bottom of the page.